chore: update Sound and Sprite classes to current UML
