--- bridge/src/main/java/com/bridge/renderHandler/factory/ISpriteFactory.java
@@ -1,16 +1,0 @@
-package com.bridge.renderHandler.factory;
-
-import com.bridge.renderHandler.sprite.*;
-
-/**
- * ISpriteFactory interface for sprite creation.
- */
-public interface ISpriteFactory {
-
-    /**
-     * Creates a new sprite.
-     *
-     * @return the sprite created
-     */
-    Sprite createSprite();
-}
