+++ bridge/src/main/java/com/bridge/renderHandler/sound/SoundBuilder.java
@@ -10,7 +10,7 @@ import java.nio.file.Path;
  */
 public class SoundBuilder {
     private IRepository<Sound> soundRepository;
-    private FileHandler fileHandler;
+    private final FileHandler fileHandler;
     private Path filePath;
 
     /**
