+++ bridge/src/main/java/com/bridge/renderHandler/sprite/attributes/Coord.java
@@ -1,4 +1,4 @@
-package com.bridge.renderHandler;
+package com.bridge.renderHandler.sprite.attributes;
 
 /**
  * Coord class that represents the position of a sprite.
